int to size_t

+++ inc/slmseg.hpp
@@ -28,18 +28,18 @@ public:
             return b + std::log1p(std::exp(a - b));
     }
 
-    void transemisi_HSLM(const std::vector<TFloat>& etavec, int NCov, int K, int NExp, int T,
+    void transemisi_HSLM(const std::vector<TFloat>& etavec, size_t NCov, size_t K, size_t NExp, size_t T,
                     std::vector<TFloat>& G, std::vector<std::vector<TFloat>>& P, std::vector<std::vector<TFloat>>& Emission);
 
-    void transemisi_SLM(const TFloat eta, int K, int NExp, int T,
+    void transemisi_SLM(const TFloat eta, size_t K, size_t NExp, size_t T,
                     std::vector<TFloat>& G, std::vector<std::vector<TFloat>>& P, std::vector<std::vector<TFloat>>& Emission);                  
                     
     void bioviterbii_HSLM(const std::vector<TFloat>& etav, const std::vector<std::vector<TFloat>>& P, 
-                    const std::vector<std::vector<TFloat>>& Emission, int  T, int  K, std::vector<unsigned int>& path, 
+                    const std::vector<std::vector<TFloat>>& Emission, size_t  T, size_t  K, std::vector<unsigned int>& path, 
                     std::vector<std::vector<unsigned int>>& psi);
     
     void bioviterbii_SLM(const std::vector<TFloat>& etav, const std::vector<std::vector<TFloat>>& P, 
-                    const std::vector<std::vector<TFloat>>& Emission, int  T, int  K, std::vector<unsigned int>& path, 
+                    const std::vector<std::vector<TFloat>>& Emission, size_t  T, size_t  K, std::vector<unsigned int>& path, 
                     std::vector<std::vector<unsigned int>>& psi); 
     
     void param_est_seq();
@@ -103,11 +103,11 @@ std::vector<unsigned int> SLMSeg<TFloat>::get_breaks(std::vector<unsigned int>& 
     if (path.empty())
         return breakpoints;
 
-    int T = path.size();
+    size_t T = path.size();
     unsigned int last_state = path[0];
     breakpoints.push_back(0);  // 起點一定是斷點
 
-    for (int t = 1; t < T; ++t) {
+    for (size_t t = 1; t < T; ++t) {
         if (path[t] != last_state) {
             breakpoints.push_back(t);  // 狀態切換時記錄斷點
             last_state = path[t];
@@ -156,10 +156,10 @@ bool SLMSeg<TFloat>::load_signal_file(const std::string& file_name, const std::s
 template<class TFloat>
 void SLMSeg<TFloat>::transemisi_HSLM(
     const std::vector<TFloat>& etavec,
-    int NCov,
-    int K,
-    int NExp,
-    int T,
+    size_t NCov,
+    size_t K,
+    size_t NExp,
+    size_t T,
     std::vector<TFloat>& G, 
     std::vector<std::vector<TFloat>>& P, 
     std::vector<std::vector<TFloat>>& Emission) 
@@ -167,9 +167,9 @@ void SLMSeg<TFloat>::transemisi_HSLM(
     TFloat PI = std::numbers::pi;
     // Step 1: GVECT (a simplify gaussion log-likelihood)
     std::vector<TFloat> gvect(K, 0.0);
-    for (int i = 0; i < K; ++i) {
+    for (size_t i = 0; i < K; ++i) {
         TFloat gsum = 0.0;
-        for (int j = 0; j < NExp; ++j) {
+        for (size_t j = 0; j < NExp; ++j) {
             TFloat diff = muk_[j][i] - mi_[j];
             gsum += - (diff * diff) / (2 * smu_[j] * smu_[j]);
         }
@@ -178,15 +178,15 @@ void SLMSeg<TFloat>::transemisi_HSLM(
 
     // Step 2: normalize GVECT (likelihood) to G (probability)
     TFloat norm = gvect[0];
-    for (int i = 1; i < K; ++i)
+    for (size_t i = 1; i < K; ++i)
         norm = elnsum(norm, gvect[i]);
-    for (int i = 0; i < K; ++i)
+    for (size_t i = 0; i < K; ++i)
         G[i] = gvect[i] - norm;
 
     // Step 3: transition matrix P [K][K][NCov]
-    for (int cov = 0; cov < NCov; ++cov) {
-        for (int i = 0; i < K; ++i) {
-            for (int j = 0; j < K; ++j) {
+    for (size_t cov = 0; cov < NCov; ++cov) {
+        for (size_t i = 0; i < K; ++i) {
+            for (size_t j = 0; j < K; ++j) {
                 if (i == j)
                     P[i][j + cov * K] = elnsum(std::log(1 - etavec[cov]), std::log(etavec[cov]) + G[j]);
                 else
@@ -196,10 +196,10 @@ void SLMSeg<TFloat>::transemisi_HSLM(
     }
 
     // Step 4: emission log-likelihood [K x T]
-    for (int k = 0; k < K; ++k) {
-        for (int t = 0; t < T; ++t) {
+    for (size_t k = 0; k < K; ++k) {
+        for (size_t t = 0; t < T; ++t) {
             TFloat val = 0.0;
-            for (int j = 0; j < NExp; ++j) {
+            for (size_t j = 0; j < NExp; ++j) {
                 TFloat diff = data_matrix[j][t] - muk_[j][k];
                 val += -0.5 * (diff * diff) / (sepsilon_[j] * sepsilon_[j]) -
                     std::log(std::sqrt(2.0 * PI) * sepsilon_[j]);
@@ -212,9 +212,9 @@ void SLMSeg<TFloat>::transemisi_HSLM(
 template<class TFloat>
 void SLMSeg<TFloat>::transemisi_SLM(
     const TFloat eta,
-    int K,
-    int NExp,
-    int T,
+    size_t K,
+    size_t NExp,
+    size_t T,
     std::vector<TFloat>& G, 
     std::vector<std::vector<TFloat>>& P, 
     std::vector<std::vector<TFloat>>& Emission) 
@@ -223,9 +223,9 @@ void SLMSeg<TFloat>::transemisi_SLM(
 
     // Step 1: GVECT
     std::vector<TFloat> gvect(K, 0.0);
-    for (int i = 0; i < K; ++i) {
+    for (size_t i = 0; i < K; ++i) {
         TFloat gsum = 0.0;
-        for (int j = 0; j < NExp; ++j) {
+        for (size_t j = 0; j < NExp; ++j) {
             TFloat diff = muk_[j][i] - mi_[j];
             gsum += - (diff * diff) / (2 * smu_[j] * smu_[j]);
         }
@@ -234,14 +234,14 @@ void SLMSeg<TFloat>::transemisi_SLM(
 
     // Step 2: Normalize GVECT to G
     TFloat norm = gvect[0];
-    for (int i = 1; i < K; ++i)
+    for (size_t i = 1; i < K; ++i)
         norm = elnsum(norm, gvect[i]);
-    for (int i = 0; i < K; ++i)
+    for (size_t i = 0; i < K; ++i)
         G[i] = gvect[i] - norm;
 
     // Step 3: Transition matrix P (K x K)
-    for (int i = 0; i < K; ++i) {
-        for (int j = 0; j < K; ++j) {
+    for (size_t i = 0; i < K; ++i) {
+        for (size_t j = 0; j < K; ++j) {
             if (i == j)
                 P[i][j] = elnsum(std::log(1.0 - eta), std::log(eta) + G[j]);
             else
@@ -250,10 +250,10 @@ void SLMSeg<TFloat>::transemisi_SLM(
     }
 
     // Step 4: Emission likelihoods (K x T)
-    for (int k = 0; k < K; ++k) {
-        for (int t = 0; t < T; ++t) {
+    for (size_t k = 0; k < K; ++k) {
+        for (size_t t = 0; t < T; ++t) {
             TFloat val = 0.0;
-            for (int j = 0; j < NExp; ++j) {
+            for (size_t j = 0; j < NExp; ++j) {
                 TFloat diff = data_matrix[j][t] - muk_[j][k];
                 val += -0.5 * (diff * diff) / (sepsilon_[j] * sepsilon_[j]) -
                        std::log(std::sqrt(2.0 * PI) * sepsilon_[j]);
@@ -268,26 +268,26 @@ void SLMSeg<TFloat>::bioviterbii_HSLM(
     const std::vector<TFloat>& etav, 
     const std::vector<std::vector<TFloat>>& P, 
     const std::vector<std::vector<TFloat>>& Emission,
-    int T, 
-    int K, 
+    size_t T, 
+    size_t K, 
     std::vector<unsigned int>& path, 
     std::vector<std::vector<unsigned int>>& psi) 
 {
     std::vector<std::vector<TFloat>> delta(K, std::vector<TFloat>(T, 0.0));
     
     // initialization
-    for (int i = 0; i < K; ++i) {
+    for (size_t i = 0; i < K; ++i) {
         delta[i][0] = etav[i] + Emission[i][0];
         psi[i][0] = 0;
     }
 
     // recursion
-    for (int t = 1; t < T; ++t) {
-        for (int j = 0; j < K; ++j) {
+    for (size_t t = 1; t < T; ++t) {
+        for (size_t j = 0; j < K; ++j) {
             TFloat nummax = delta[0][t - 1] + P[0][j + K * (t - 1)]; // P[0][j][t]
             unsigned int ind = 0;
 
-            for (int i = 1; i < K; ++i) {
+            for (size_t i = 1; i < K; ++i) {
                 TFloat score = delta[i][t - 1] + P[i][j + K * (t - 1)]; // P[i][j][t-1]
                 if (score > nummax) {
                     nummax = score;
@@ -303,7 +303,7 @@ void SLMSeg<TFloat>::bioviterbii_HSLM(
     // termination
     TFloat maxval = delta[0][T - 1];
     unsigned int ind = 0;
-    for (int i = 1; i < K; ++i) {
+    for (size_t i = 1; i < K; ++i) {
         if (delta[i][T - 1] > maxval) {
             maxval = delta[i][T - 1];
             ind = i;
@@ -322,26 +322,26 @@ void SLMSeg<TFloat>::bioviterbii_SLM(
     const std::vector<TFloat>& etav, 
     const std::vector<std::vector<TFloat>>& P, 
     const std::vector<std::vector<TFloat>>& Emission,
-    int T, 
-    int K, 
+    size_t T, 
+    size_t K, 
     std::vector<unsigned int>& path, 
     std::vector<std::vector<unsigned int>>& psi) 
 {
     std::vector<std::vector<TFloat>> delta(K, std::vector<TFloat>(T, 0.0));
 
     // Initialization
-    for (int i = 0; i < K; ++i) {
+    for (size_t i = 0; i < K; ++i) {
         delta[i][0] = etav[i] + Emission[i][0];
         psi[i][0] = 0;
     }
 
     // Recursion
-    for (int t = 1; t < T; ++t) {
-        for (int j = 0; j < K; ++j) {
+    for (size_t t = 1; t < T; ++t) {
+        for (size_t j = 0; j < K; ++j) {
             TFloat max_score = delta[0][t - 1] + P[0][j];
             unsigned int max_index = 0;
 
-            for (int i = 1; i < K; ++i) {
+            for (size_t i = 1; i < K; ++i) {
                 TFloat score = delta[i][t - 1] + P[i][j];
                 if (score > max_score) {
                     max_score = score;
@@ -357,7 +357,7 @@ void SLMSeg<TFloat>::bioviterbii_SLM(
     // Termination
     TFloat maxval = delta[0][T - 1];
     unsigned int maxind = 0;
-    for (int i = 1; i < K; ++i) {
+    for (size_t i = 1; i < K; ++i) {
         if (delta[i][T - 1] > maxval) {
             maxval = delta[i][T - 1];
             maxind = i;
@@ -366,7 +366,7 @@ void SLMSeg<TFloat>::bioviterbii_SLM(
 
     // Traceback
     path[T - 1] = maxind;
-    for (int t = T - 2; t >= 0; --t) {
+    for (size_t t = T - 2; t >= 0; --t) {
         path[t] = psi[path[t + 1]][t + 1];
     }
 }
@@ -377,11 +377,11 @@ void SLMSeg<TFloat>::param_est_seq()
     // smu_: Between-segment variation
     // sepsilon_: Within-segment noise
     data_matrix.emplace_back(signal_data_); // log2R
-    int Nexp = data_matrix.size();
+    size_t Nexp = data_matrix.size();
 
     std::vector<TFloat> sigmax(Nexp);
 
-    for (int i = 0; i < Nexp; ++i) {
+    for (size_t i = 0; i < Nexp; ++i) {
         // for single sample
         std::vector<TFloat>& row = data_matrix[i]; 
         std::vector<TFloat> centered(row.size());
@@ -390,7 +390,7 @@ void SLMSeg<TFloat>::param_est_seq()
         std::nth_element(tmp.begin(), tmp.begin() + tmp.size() / 2, tmp.end()); // sort nth element
         TFloat median = tmp[tmp.size() / 2]; 
         // caculate absolute deviation from the median
-        for (int j = 0; j < row.size(); ++j) {
+        for (size_t j = 0; j < row.size(); ++j) {
             centered[j] = std::abs(row[j] - median);
         }
         // caculate mad (median absolute deviation)
@@ -412,7 +412,7 @@ void SLMSeg<TFloat>::param_est_seq()
 template<class TFloat>
 void SLMSeg<TFloat>::muk_est() 
 {
-    int Nexp = data_matrix.size();
+    size_t Nexp = data_matrix.size();
 
     if (Nexp == 1) {
         std::vector<TFloat> temp;
@@ -433,12 +433,12 @@ template<class TFloat>
 void SLMSeg<TFloat>::joint_seg()
 {
     // 檢查資料
-    int NExp = data_matrix.size();
+    size_t NExp = data_matrix.size();
     if (NExp == 0 || data_matrix[0].empty())
         throw std::invalid_argument("Data matrix is empty.");
 
-    int T = data_matrix[0].size();           // 序列長度
-    int K0 = muk_[0].size();                 // 狀態數（μ的組合數）
+    size_t T = data_matrix[0].size();           // 序列長度
+    size_t K0 = muk_[0].size();                 // 狀態數（μ的組合數）
     std::vector<TFloat> etav(K0, std::log(1.0 / K0));  // etav = log(1/K0)
 
     // Step 1: 建立轉移矩陣與機率表
@@ -468,22 +468,22 @@ void SLMSeg<TFloat>::joint_seg_in()
     // CovPosNorm = CovPos / stepeta(dnorm)
     // Pr = etavec = eta + (1 - eta) * exp(log(eta) / CovPosNorm) 
     std::vector<TFloat> etavec(pos_data_.size() - 1);
-    for (int i = 0; i < etavec.size(); ++i) {
+    for (size_t i = 0; i < etavec.size(); ++i) {
        TFloat cov_pos = static_cast<TFloat>(pos_data_[i + 1]) - static_cast<TFloat>(pos_data_[i]);
        TFloat cov_pos_norm = cov_pos / stepeta_;
        etavec[i] = eta_ + (1.0 - eta_) * std::exp(std::log(eta_) / cov_pos_norm);
     }
     // 樣本數 NExp = sample size
-    int NExp = data_matrix.size();
+    size_t NExp = data_matrix.size();
     if (NExp == 0 || data_matrix[0].empty())
     throw std::invalid_argument("Data matrix is empty.");
 
     // bins pair NCov = length(etavec)
-    int NCov = etavec.size();
+    size_t NCov = etavec.size();
     // 狀態數（候選的 mean 組合 μ_k）K0 = ncol(muk)
-    int K0 = muk_[0].size();
+    size_t K0 = muk_[0].size();
     // bins number
-    int T = data_matrix[0].size();
+    size_t T = data_matrix[0].size();
     // etav = log(rep(1, K0) * (1/K0))
     std::vector<TFloat> etav(K0, std::log(1.0 / K0));
 
@@ -513,7 +513,7 @@ void SLMSeg<TFloat>::joint_seg_in()
 template<class TFloat>
 void SLMSeg<TFloat>::filter_seg() {
     total_pred_break_filtered_.clear(); 
-    int n = total_pred_break_.size();
+    size_t n = total_pred_break_.size();
     if (n == 0) return;
 
     std::vector<unsigned int> indF;
@@ -543,12 +543,12 @@ void SLMSeg<TFloat>::filter_seg() {
 template<class TFloat>
 void SLMSeg<TFloat>::seg_results()
 {
-    int NExp = data_matrix.size();
-    int T = data_matrix[0].size();
+    size_t NExp = data_matrix.size();
+    size_t T = data_matrix[0].size();
 
-    for (int j = 0; j < NExp; ++j) {
+    for (size_t j = 0; j < NExp; ++j) {
         std::vector<TFloat> s(T, 0.0);
-        for (int i = 0; i + 1 < total_pred_break_filtered_.size(); ++i) {
+        for (size_t i = 0; i + 1 < total_pred_break_filtered_.size(); ++i) {
             unsigned int start = total_pred_break_filtered_[i];
             unsigned int end = total_pred_break_filtered_[i + 1];
 
@@ -556,7 +556,7 @@ void SLMSeg<TFloat>::seg_results()
             std::nth_element(segment.begin(), segment.begin() + segment.size() / 2, segment.end());
             TFloat median = segment[segment.size() / 2];
 
-            for (int t = start; t < end; ++t)
+            for (size_t t = start; t < end; ++t)
                 s[t] = median;
         }
         data_seg_.push_back(std::move(s));
